Output.txt
Output.txt contains the output of the data obtained through extract.py, which extracts the data from the database using threads, each thread is independent of the other, as the command to extract data from the database is divided into 4 threads(depends on the no of cores of the system) each thread deals with a part of the database. Here in my project as i have a data set of more than 1000 cities with coordinates and name. Here I have distributed the work of fetching the data to four threads thus decreasing the execution time.

Every thread fetch the data of the quarter of the database and when any of the threads gets a data it stops all the process and returns a value.ie, the coordinates of the cities in between the shortest route.